--- packages/wxt/src/client/content-scripts/index.ts
@@ -1,2 +1,0 @@
-export * from './content-script-context';
-export * from './ui';
