--- packages/wxt/src/client/content-scripts/ui/index.ts
@@ -1,4 +1,0 @@
-export type * from './types';
-export * from './create-iframe-ui';
-export * from './create-integrated-ui';
-export * from './create-shadow-root-ui';
