--- packages/wxt/src/client/index.ts
@@ -1,8 +1,0 @@
-/**
- * Any runtime APIs that use the web extension APIs.
- *
- * @module wxt/client
- */
-export * from './content-scripts';
-export * from './app-config';
-export * from './inject-script';
