--- packages/wxt/src/sandbox/index.ts
@@ -1,11 +1,0 @@
-/**
- * Any runtime APIs that don't use the web extension APIs.
- *
- * @module wxt/sandbox
- */
-export * from './define-unlisted-script';
-export * from './define-background';
-export * from './define-content-script';
-export * from './define-wxt-plugin';
-export * from './define-app-config';
-export * from './match-patterns';
