+++ packages/wxt/src/utils/README.md
@@ -1,0 +1,1 @@
+This folder is for public utils, not internal utils. Put generic helpers and other utils in the core/utils folder.
