+++ packages/wxt/src/utils/__tests__/content-script-context.test.ts
