+++ packages/wxt/src/utils/__tests__/define-background.test.ts
