+++ packages/wxt/src/utils/__tests__/define-content-script.test.ts
