+++ packages/wxt/src/utils/__tests__/define-unlisted-script.test.ts
