+++ packages/wxt/src/utils/app-config.ts
