+++ packages/wxt/src/utils/content-script-context.ts
