+++ packages/wxt/src/utils/content-script-ui/integrated.ts
