+++ packages/wxt/src/utils/content-script-ui/shadow-root.ts
