+++ packages/wxt/src/utils/content-script-ui/shared.ts
