+++ packages/wxt/src/utils/define-app-config.ts
