+++ packages/wxt/src/utils/define-background.ts
