+++ packages/wxt/src/utils/define-content-script.ts
