+++ packages/wxt/src/utils/define-unlisted-script.ts
