+++ packages/wxt/src/utils/define-wxt-plugin.ts
