+++ packages/wxt/src/utils/inject-script.ts
