+++ packages/wxt/src/utils/internal/custom-events.ts
