+++ packages/wxt/src/utils/internal/location-watcher.ts
