+++ packages/wxt/src/utils/internal/logger.ts
