+++ packages/wxt/src/utils/match-patterns.ts
