+++ packages/wxt/src/virtual/content-script-main-world-entrypoint.ts
@@ -1,5 +1,5 @@
 import definition from 'virtual:user-content-script-main-world-entrypoint';
-import { logger } from '../sandbox/utils/logger';
+import { logger } from '../utils/internal/logger';
 import { initPlugins } from 'virtual:wxt-plugins';
 
 const result = (async () => {
